test, refactor : Repository에 RacingCar 저장 후 자동차 전진
1. refactor :
- OperatorControllerTest 클래스의 repository에_RacingCar_저장(), repository에_RacingCar_저장_후_자동차_전진_또는_정지()를 RacingCarServiceTest 클래스로 이동.
- RacingCarRepository를 Singleton Holder 방식으로 생성하도록 개선

2. test :
> RacingCarServiceTest에서 RacingCar_저장(), repository에_RacingCar_저장_후_자동차_전진_또는_정지() test case 테스트
