+++ src/main/java/racingcar/domain/racingcar/RacingCarRepository.java
@@ -8,8 +8,18 @@ import racingcar.dto.RacingCarDto;
 public class RacingCarRepository {
 	private Map<String, RacingCar> racingCarMap = new HashMap<>();
 
-	public RacingCarRepository() {
+	// start: Singleton Holder
+	private RacingCarRepository() {
 	}
+
+	private static class InnerRacingCarRepository {
+		private static final RacingCarRepository instance = new RacingCarRepository();
+	}
+
+	public static RacingCarRepository getInstance() {
+		return InnerRacingCarRepository.instance;
+	}
+	// end: Singleton Holder
 
 	public Map<String, RacingCar> getRacingCarMap() {
 		return racingCarMap;
