+++ src/test/java/racingcar/controller/OperatorControllerTest.java
@@ -10,13 +10,10 @@ import org.junit.jupiter.api.Test;
 import camp.nextstep.edu.missionutils.test.NsTest;
 import racingcar.domain.enumtype.InterfaceMsg;
 import racingcar.domain.enumtype.ValidationMsg;
-import racingcar.domain.racingcar.RacingCarRepository;
-import racingcar.dto.RacingCarDto;
 import racingcar.service.ValidatorServiceTest;
 
 public class OperatorControllerTest extends NsTest {
 	private ValidatorServiceTest validatorServiceTest = ValidatorServiceTest.getInstance();
-	private RacingCarDto racingCarDto;
 
 	@Test
 	void 자동차_이름_string_null_입력_체크() {
@@ -109,26 +106,6 @@ public class OperatorControllerTest extends NsTest {
 				System.out.println(validationMsg.getValue()); throw new IllegalArgumentException();
 			}
 		});
-	}
-
-	@DisplayName("OperatorControllerTest.repository에_RacingCar_저장()")
-	@Test
-	void repository에_RacingCar_저장() {
-		RacingCarDto racingCarDto = new RacingCarDto("pobi2");
-		RacingCarRepository racingCarRepository = new RacingCarRepository();
-		racingCarRepository.initSaveRacingCar(racingCarDto);
-		assertEquals(racingCarDto.getCarName(), racingCarRepository.getRacingCarByName("pobi2").getCarName());
-	}
-
-	@DisplayName("OperatorControllerTest.repository에_RacingCar_저장_후_자동차_전진_또는_정지()")
-	@Test
-	void repository에_RacingCar_저장_후_자동차_전진_또는_정지() {
-		racingCarDto = RacingCarDto.builder().carName("pobi2").build();
-		RacingCarRepository racingCarRepository = new RacingCarRepository();
-		racingCarRepository.initSaveRacingCar(racingCarDto);
-		System.out.println(racingCarRepository.getRacingCarByName(racingCarDto.getCarName()).getCarPosition());
-		racingCarRepository.movingForwardByName(racingCarDto);
-		System.out.println(racingCarRepository.getRacingCarByName(racingCarDto.getCarName()).getCarPosition());
 	}
 
 	@Override
